+++ biomes/river.py
@@ -1,16 +1,16 @@
-from interfaces import IAquatic
+from interfaces import IsAquatic
 from interfaces import Identifiable
 from interfaces import IContainsAnimals
 from interfaces import IContainsPlants
-from animals import RiverDolphin
+# from animals import RiverDolphin
 
 
 class River(IContainsAnimals, IContainsPlants, Identifiable):
 
     def __init__(self):
-      IContainsAnimals.__init__(self)
-      IContainsPlants.__init__(self)
-      Identifiable.__init__(self)
+        IContainsAnimals.__init__(self)
+        IContainsPlants.__init__(self)
+        Identifiable.__init__(self)
 
     def add_animal(self, animal):
         try:
