+++ index.py
@@ -2,7 +2,7 @@
 
 import os
 from arboretum import Arboretum
-from actions.annex import annex_habitat
+from actions.annex import annex_biome
 from actions.release_animal import release_animal
 from actions.report import build_facility_report
 from actions.cultivate_plant_menu import cultivate_plant_menu
@@ -12,8 +12,8 @@ keahua = Arboretum("Keahua Arboretum", "123 Paukauila Lane")
 
 def build_menu():
     os.system('cls' if os.name == 'nt' else 'clear')
-    print("1. Annex Habitat")
-    print("2. Release Animal into Habitat")
+    print("1. Annex Biome")
+    print("2. Release Animal into Biome")
     print("3. Feed Animal")
     print("4. Cultivate New Plant")
     print("5. Display Facility Report")
@@ -28,7 +28,7 @@ def main_menu():
     choice = input(">> ")
 
     if choice == "1":
-        annex_habitat(keahua)
+        annex_biome(keahua)
 
     if choice == "2":
         release_animal(keahua)
