+++ interfaces/__init__.py
@@ -1,9 +1,9 @@
-from .terrestrial import ITerrestrial
-from .animal.freshwater import IFreshwater
-from interfaces.aquatic import IAquatic
-from .flying import IFlying
-from .walking import IWalking
-from .swimming import ISwimming
+from .animals.terrestrial import IsTerrestrial
+from .animals.freshwater import IsFreshwater
+from interfaces.animals.aquatic import IsAquatic
+from .animals.actions.flying import IsFlying
+from .animals.actions.walking import IsWalking
+from .animals.actions.swimming import IsSwimming
 from .identifiable import Identifiable
-from .habitat.contains_animals import IContainsAnimals
-from .habitat.contains_plants import IContainsPlants
+from .habitats.contains_animals import IContainsAnimals
+from .habitats.contains_plants import IContainsPlants
