+++ interfaces/animals/__init__.py
@@ -1,0 +1,4 @@
+from .freshwater import IsFreshwater
+from .saltwater import IsSaltwater
+from .terrestrial import IsTerrestrial
+from .aquatic import IsAquatic
