+++ interfaces/animals/aquatic.py
@@ -1,4 +1,4 @@
-class IAquatic:
+class IsAquatic:
 
     def __init__(self):
         self.aquatic = True
