+++ interfaces/animals/freshwater.py
@@ -1,7 +1,7 @@
-from interfaces.aquatic import IAquatic
+from interfaces.animals.aquatic import IsAquatic
 # Error 1 solving Circular Import due to import path being without direction to aquatic. Example: from aquatic import IAquatic
 # Solution from interfaces.aquatic import IAquatic
-class IFreshwater(IAquatic):
+class IsFreshwater(IsAquatic):
 
     def __init__(self):
         super().__init__()
