+++ interfaces/animals/terrestrial.py
@@ -1,4 +1,4 @@
-class ITerrestrial:
+class IsTerrestrial:
 
     def __init__(self):
         self.terrestrial = True
